requirements.txt
aiosmtplib
alembic
bcrypt
email_validator
fastapi
fastapi-mail
Jinja2
passlib
psycopg2-binary
pydantic
PyJWT
python-dotenv
python-jose
python-multipart
SQLAlchemy
uvicorn
authlib
itsdangerous
requests
aioboto3
filetype